+++ solutions/e-commerce-springcloud-microservices/cart-service/src/main/java/com/htsat/cart/service/IUserService.java
@@ -2,5 +2,5 @@ package com.htsat.cart.service;
 
 public interface IUserService {
 
-    boolean checkUserAvailable(int userId);
+    boolean checkUserAvailable(Long userId);
 }
