+++ solutions/e-commerce-springcloud-microservices/order-service/src/main/java/com/htsat/order/service/IAddressService.java
@@ -2,5 +2,5 @@ package com.htsat.order.service;
 
 public interface IAddressService {
 
-    boolean checkAddressAvailable(int userId, int addressId);
+    boolean checkAddressAvailable(Long userId, Long addressId);
 }
